requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langgraph
langchain-text-splitters

google-genai

qdrant-client

pydantic
pydantic-settings
python-multipart

tiktoken

pypdf